Move the OutstationTaskAdapter to its own module. It'll be useful for TLS as well eventually
