+++ dnp3/src/tcp/outstation.rs
@@ -3,104 +3,13 @@ use tracing::Instrument;
 use crate::app::Shutdown;
 use crate::link::LinkErrorMode;
 use crate::outstation::database::EventBufferConfig;
-use crate::outstation::session::RunError;
 use crate::outstation::task::OutstationTask;
 use crate::outstation::OutstationHandle;
 use crate::outstation::*;
 use crate::tcp::{AddressFilter, FilterError};
-use crate::util::channel::{request_channel, Receiver, Sender};
-use crate::util::phys::PhysLayer;
+use crate::util::channel::Sender;
 
-/// message that gets sent to OutstationTaskAdapter when
-/// it needs to switch to a new session
-#[derive(Debug)]
-pub(crate) struct NewSession {
-    pub(crate) id: u64,
-    pub(crate) phys: PhysLayer,
-}
-
-impl NewSession {
-    pub(crate) fn new(id: u64, phys: PhysLayer) -> Self {
-        Self { id, phys }
-    }
-}
-
-/// adapts an OutstationTask to something that can listen for new
-/// connections on a channel and shutdown an existing session
-struct OutstationTaskAdapter {
-    receiver: Receiver<NewSession>,
-    task: OutstationTask,
-}
-
-impl OutstationTaskAdapter {
-    fn create(task: OutstationTask) -> (Self, Sender<NewSession>) {
-        let (tx, rx) = request_channel();
-
-        (Self { receiver: rx, task }, tx)
-    }
-
-    async fn wait_for_session(&mut self) -> Result<NewSession, Shutdown> {
-        loop {
-            crate::tokio::select! {
-                session = self.receiver.receive() => {
-                    return session;
-                }
-                ret = self.task.process_messages() => {
-                    ret?
-                }
-            }
-        }
-    }
-
-    async fn run_one_session(&mut self, io: &mut PhysLayer) -> Result<NewSession, RunError> {
-        crate::tokio::select! {
-            res = self.task.run(io) => {
-                Err(res)
-            }
-            x = self.receiver.receive() => {
-                Ok(x?)
-            }
-        }
-    }
-
-    async fn run(&mut self) -> Result<(), Shutdown> {
-        let mut session = None;
-
-        loop {
-            match session.take() {
-                None => {
-                    session.replace(self.wait_for_session().await?);
-                }
-                Some(mut s) => {
-                    let id = s.id;
-
-                    let result = self
-                        .run_one_session(&mut s.phys)
-                        .instrument(tracing::info_span!("Session", "id" = id))
-                        .await;
-
-                    match result {
-                        Ok(new_session) => {
-                            // TODO reset the task
-                            tracing::warn!(
-                                "closing session {} for new session {}",
-                                id,
-                                new_session.id
-                            );
-                            // go to next iteration with a new session
-                            session.replace(new_session);
-                        }
-                        Err(RunError::Link(err)) => {
-                            // TODO reset the task
-                            tracing::warn!("Session error: {}", err);
-                        }
-                        Err(RunError::Shutdown) => return Err(Shutdown),
-                    }
-                }
-            }
-        }
-    }
-}
+use crate::outstation::adapter::{NewSession, OutstationTaskAdapter};
 
 struct OutstationInfo {
     filter: AddressFilter,
@@ -234,10 +143,10 @@ impl TcpServer {
 
         crate::tokio::select! {
              _ = self.accept_loop(listener) => {
-
+                // if the accept loop shuts down we exit
              }
              _ = rx => {
-
+                // if we get the message or shutdown we exit
              }
         }
